Refactor app settings
